rqe_iterators_bencher: split bencher into numeric, term, and wildcard submodules
We did the same split for Rust code, tests and ffi.
Will help reducing conflicts.
